requirements.txt
requests==2.21.0
pandas==0.24.2
PyJWT==1.7.1
python-dateutil==2.8.0
